Switch all validation errors to use the ValidationError class
For consistency
