+++ app/jobs/shelve_job.rb
@@ -42,6 +42,6 @@ class ShelveJob < ApplicationJob
 
   def validator_for?(item)
     model = Cocina::Mapper.build(item)
-    ValidateDarkService.new(model)
+    Cocina::ValidateDarkService.new(model)
   end
 end
