+++ spec/jobs/shelve_job_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe ShelveJob, type: :job do
   let(:druid) { 'druid:mk420bs7601' }
   let(:result) { create(:background_job_result) }
   let(:item) { instance_double(Dor::Item) }
-  let(:validator) { instance_double(ValidateDarkService, valid?: valid, invalid_filenames: invalid_filenames) }
+  let(:validator) { instance_double(Cocina::ValidateDarkService, valid?: valid, invalid_filenames: invalid_filenames) }
   let(:valid) { true }
   let(:invalid_filenames) { [] }
 
@@ -17,7 +17,7 @@ RSpec.describe ShelveJob, type: :job do
     allow(result).to receive(:processing!)
     allow(EventFactory).to receive(:create)
     allow(Cocina::Mapper).to receive(:build)
-    allow(ValidateDarkService).to receive(:new).and_return(validator)
+    allow(Cocina::ValidateDarkService).to receive(:new).and_return(validator)
   end
 
   context 'with no errors' do
